+++ src/types/index.ts
@@ -7,7 +7,7 @@ export class Coverage {
     this.covered = Number(covered);
     this.percentual =
       this.total == 0
-        ? undefined
+        ? 1.0
         : parseFloat((this.covered / this.total).toFixed(4));
   }
 }